services/event-processor/requirements.txt
kafka-python==2.0.2
qdrant-client==1.7.3
sentence-transformers==2.3.1
